Домашнее задание номер 5
