+++ js/decription-photo.js
@@ -1,0 +1,17 @@
+import {getRandomArrayElement, getRandomNumber, getRandomComment, description} from 'util.js';
+
+let photoID = 0;
+
+const descriptionPhoto = function() {
+  ++photoID;
+  return {
+    id: photoID,
+    url: `photos/${photoID}.jpg`,
+    description:  getRandomArrayElement(description),
+    likes: getRandomNumber(15, 200),
+    comments: Array.from({length:getRandomNumber(1, 15)}, getRandomComment),
+  };
+};
+
+// eslint-disable-next-line
+const photoContent = Array.from({length: 25}, descriptionPhoto);
